# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Hinter Factory ML API
  version: 0.1.0
  description: >-
    Every scoped endpoint (documents, tags, labeling functions, LF runs,
    gold labels, probabilistic labels, evaluation) requires a `project_id`
    query parameter (or, for multipart uploads, a `project_id` form field).
    There is no implicit fallback project; calls that omit it return 400.
    Use `GET /v1/projects` to discover available projects and
    `POST /v1/projects` to create one. The web UI threads `project_id`
    through every request automatically via its project context.
servers:
  - url: http://localhost:8000
...
paths:
...
  /v1/tags:
    get:
      tags: [tags]
      operationId: listTags
      parameters:
        - in: query
          name: project_id
          required: false
          schema: { type: string }
          description: >-
            Project to scope the listing to. Required at the runtime / API level
            (the server returns 400 when it's missing); marked optional in the
            schema only so the web client's openapi-fetch middleware can inject
            it transparently from the active project header.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
        "400":
          description: project_id missing
    post:
      tags: [tags]
      operationId: createTag
      parameters:
        - in: query
          name: project_id
          required: false
          schema: { type: string }
          description: >-
            Project to create the tag in. Required at the runtime / API level
            (the server returns 400 when it's missing); marked optional in the
            schema only so the web client's openapi-fetch middleware can inject
            it from the active project header. Wins over a `project_id` field
            in the request body if both are present.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
...
components:
...
  schemas:
...
    Tag:
      type: object
      required: [id, name, taxonomy_version, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        taxonomy_version: { type: string }
        created_at: { type: string, format: date-time }
    TagCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        taxonomy_version: { type: string, default: "v1" }
...